Copy over conf files from old heroku-php-buildpack repo
When we switched to this buildpack, we just wanted to gain grunt
compilation, but we ended up with a different apache/php config. not
sure what all of these do / if they affect anything, but let's keep them
in order to maintain the same server conf. check out the other repo
(southern-experience/heroku-php-buildpack) for individual commits on
these changes